extract `TimerWrapper`
