+++ src/api/communication.ts
@@ -13,7 +13,7 @@
 import { APPLICATION_NAME } from './env.ts';
 import { ERRORS_IGNORED } from './const.ts';
 import type { TMetrics } from '../api-monitor-cs-main.ts';
-import { ETimerType } from '../wrapper/Wrapper.ts';
+import { ETimerType } from '../wrapper/TimerWrapper.ts';
 import type { TSettings } from './settings.ts';
 
 let port: chrome.runtime.Port | null = null;
